[Part 1] Rename Icd Management cluster files.
