+++ src/app/clusters/icd-management-server/IcdManagementCluster.cpp
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-#include "icd-management-server.h"
+#include "IcdManagementCluster.h"
 
 #include <access/AccessControl.h>
 #include <access/Privilege.h>
