+++ src/app/clusters/icd-management-server/app_config_dependent_sources.cmake
@@ -16,8 +16,8 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/icd-management-server.cpp"
-    "${CLUSTER_DIR}/icd-management-server.h"
+    "${CLUSTER_DIR}/IcdManagementCluste.cpp"
+    "${CLUSTER_DIR}/IcdManagementCluste.h"
 )
 
 # TODO(#32321): Remove after issue is resolved
