+++ src/app/clusters/icd-management-server/app_config_dependent_sources.gni
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "icd-management-server.cpp",
-  "icd-management-server.h",
+  "IcdManagementCluster.cpp",
+  "IcdManagementCluster.h",
 ]
